requirements.txt
-i https://pypi.org/simple/
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
django-cors-headers==3.2.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-templated-mail==1.1.1
django==2.2.10
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.10.3
djoser==2.0.3
drf-writable-nested==0.5.3
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
pycparser==2.19
pyjwt==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0
drf-yasg==1.21.5
